Create Voice-Countdown.ps1
I've used this script multiple times with my boys when we play Nerf wars.  The timer function keeps track of a 5 minute match and gives you a bearing of how much time is left.  I run the script in the background of some dubstep music.  It makes for some epic play times!